🚧 ログイン状況に応じて自動入力するロジックをhookに切り出し
