+++ web/src/components/championship/Form/HostInfoStep.tsx
@@ -1,7 +1,6 @@
 import { Box, Input, Textarea } from "@chakra-ui/react"
-import { FC, useEffect, useState } from "react"
-import { getDoc, profileRef } from "~/firebase"
-import { useAuthState } from "~/hooks/useAuthState"
+import { FC, useState } from "react"
+import { useUserDataEffect } from "~/hooks/useUserDataEffect"
 import { Championship } from "~/shared/firebase/firestore/scheme/championship"
 import { FormField } from "./FormField"
 import { FormStep } from "./FormStep"
@@ -61,24 +60,11 @@ export const useHostInfoStep = ({ defaultValue }: HostInfoStepTypes["hookOptions
   const [hostUid, setHostUid] = useState(defaultValue.host_uid)
   const [hostName, setHostName] = useState(defaultValue.host_name)
   const [hostContact, setHostContact] = useState(defaultValue.host_contact)
-  const { user, loading } = useAuthState()
-  useEffect(() => {
-    if (loading || !user) return
-    (async () => {
-      // useAuthStateのprofileを使うとprofileをloading中なのにprofileLoadingがfalseになってしまうため手動で取得
-      const snapshot = await getDoc(profileRef(user.uid))
-      const profile = snapshot.data()
-      setHostUid(user.uid)
-      if (profile) {
-        // profile
-        setHostName(profile.name)
-        setHostContact(profile.contact)
-      } else {
-        if (user.displayName !== null) setHostName(user.displayName)
-        if (user.email !== null && user.emailVerified) setHostContact(user.email)
-      }
-    })()
-  }, [loading, user])
+  useUserDataEffect(userData => {
+    setHostUid(userData.uid)
+    if (userData.name !== undefined) setHostName(userData.name)
+    if (userData.contact !== undefined) setHostName(userData.contact)
+  })
   const fields: HostInfoStepFields = {
     host_uid: controlledFormFieldOf(
       hostUid, setHostUid,
